Moved precompiled contracts for alt_btn128 curve operations from addresses 5 and 6 to 6 and 7
