+++ lib/index.js
@@ -11,8 +11,8 @@ const num01 = require('./precompiled/01-ecrecover.js')
 const num02 = require('./precompiled/02-sha256.js')
 const num03 = require('./precompiled/03-ripemd160.js')
 const num04 = require('./precompiled/04-identity.js')
-const num05 = require('./precompiled/05-altbn128_add.js')
-const num06 = require('./precompiled/06-altbn128_mul.js')
+const num06 = require('./precompiled/06-altbn128_add.js')
+const num07 = require('./precompiled/07-altbn128_mul.js')
 
 module.exports = VM
 
@@ -48,11 +48,11 @@ function VM (opts = {}) {
   this._precompiled['0000000000000000000000000000000000000002'] = num02
   this._precompiled['0000000000000000000000000000000000000003'] = num03
   this._precompiled['0000000000000000000000000000000000000004'] = num04
-  this._precompiled['0000000000000000000000000000000000000005'] = num05
   this._precompiled['0000000000000000000000000000000000000006'] = num06
+  this._precompiled['0000000000000000000000000000000000000007'] = num07
 
   if (this.opts.activatePrecompiles) {
-    for (var i = 1; i <= 6; i++) {
+    for (var i = 1; i <= 7; i++) {
       this.trie.put(new BN(i).toArrayLike(Buffer, 'be', 20), new Account().serialize())
     }
   }
